refactor: rename repository follow go convention
