+++ internal/server/repository.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
-	"github.com/PitiNarak/condormhub-backend/internal/repositories"
+	repository1 "github.com/PitiNarak/condormhub-backend/internal/repository"
 )
 
 type repository struct {
@@ -14,11 +14,11 @@ type repository struct {
 }
 
 func (s *Server) initRepository() {
-	user := repositories.NewUserRepo(s.db)
-	dorm := repositories.NewDormRepository(s.db)
-	leasingHistory := repositories.NewLeasingHistoryRepository(s.db)
-	order := repositories.NewOrderRepository(s.db)
-	tsx := repositories.NewTransactionRepository(s.db)
+	user := repository1.NewUserRepo(s.db)
+	dorm := repository1.NewDormRepository(s.db)
+	leasingHistory := repository1.NewLeasingHistoryRepository(s.db)
+	order := repository1.NewOrderRepository(s.db)
+	tsx := repository1.NewTransactionRepository(s.db)
 
 	s.repository = &repository{
 		user:           user,
